Lose experience and bug fixing
Created a more detailed lose experience including lose message, game flow for restarting level or returning to the main menu, stopping the game and preventing the player from clicking on the game. Also fixed three issues, two relating to null reference exceptions and one relating to an unassigned variable in the inspector.